+++ plugins/Average Primary Tip Angle.js
@@ -1,6 +1,6 @@
 const group = "Plant Measurements";
 const name = "Average Primary Tip Angle";
-const id = 'primTipAngle';
+const id = 'avgTipPrimaryAngle';
 const description = "Mean angle in degrees of all primary root tips relative to vertical, per plant";
 
 const plugin = (rsmlJson, polylines, utils) => {